add is_livechat_interactive in context
