+++ src/components/custom/other-platforms.js
@@ -290,7 +290,7 @@ export const NavResources = () => (
     <Flex wrap="wrap" jc="flex-start" max_width="42rem">
         <CardLink title={localize('Help centre')} to="/help-centre/" />
         <CardLink title={localize('Payment methods')} to="/payment-methods/" />
-        <CardLink title={localize('Community')} to={community_url} target="_blank" />
+        <CardLink title={localize('Community')} external to={community_url} target="_blank" />
         {/* TODO: add this when blog is ready */}
         {/* <CardLink title={localize('Blog')} to="/blog/" /> */}
     </Flex>
