+++ src/pages/contact-us/_contact-ways.js
@@ -36,7 +36,7 @@ const contactways = [
         text: <Localize translate_text="Our Deriv support community can help you find answers." />,
         image: 'community',
         button: (
-            <StyledLinkButton secondary="true" to={community_url} target="_blank">
+            <StyledLinkButton secondary="true" to={community_url} external target="_blank">
                 {localize('Ask the community')}
             </StyledLinkButton>
         ),
