+++ src/pages/regulatory/index.js
@@ -2,6 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import EUgrid from './_eu-grid'
 import DocumentAccordion from './_document_accordion'
+import FinancialCommission from './_financial_commission'
 import Layout from 'components/layout/layout'
 import { Header, Text, LinkText, Divider } from 'components/elements'
 import {
@@ -14,7 +15,6 @@ import {
     Show,
 } from 'components/containers'
 import { localize, WithIntl, Localize, LocalizedLink } from 'components/localization'
-import { isBrowser, deriv_app_url } from 'common/utility'
 // Icons
 import EU from 'images/svg/europe-map.svg'
 import Vanuatu from 'images/common/regulatory/vanuatu.png'
@@ -93,25 +93,11 @@ const ResponsiveHeader = styled(StyledHeader)`
     }
 `
 
-const TextLink = styled(LinkText).attrs({ as: 'span' })``
-
 const ExternalBoldLink = styled(LocalizedLink)`
     font-weight: bold;
     color: ${(props) => (props.color ? `var(--color-${props.color})` : '')};
 `
 const Regulatory = () => {
-    const LC_API = (isBrowser() && window.LC_API) || {}
-    const [is_livechat_interactive, setLiveChatInteractive] = React.useState(false)
-
-    React.useEffect(() => {
-        if (isBrowser()) {
-            window.scrollTo(0, 0)
-            window.LiveChatWidget.on('ready', () => {
-                setLiveChatInteractive(true)
-            })
-        }
-    }, [])
-
     return (
         <Layout>
             <SEO
@@ -498,50 +484,7 @@ const Regulatory = () => {
                                     ]}
                                 />
                             </Text>
-                            {is_livechat_interactive && (
-                                <div>
-                                    <Show.Eu>
-                                        <Text mt="2rem" max_width="58.8rem">
-                                            <Localize
-                                                translate_text="For fair resolution of any complaints, please speak to us using <0>chat</0>. To learn more, see our <1>complaints policy</1>."
-                                                components={[
-                                                    <TextLink
-                                                        key={0}
-                                                        color="red"
-                                                        className="gtm-deriv-livechat"
-                                                        onClick={() => {
-                                                            LC_API.open_chat_window()
-                                                        }}
-                                                    />,
-                                                    <LinkText
-                                                        key={0}
-                                                        color="red"
-                                                        target="_blank"
-                                                        rel="noopener noreferrer"
-                                                        href={`${deriv_app_url}/complaints-policy`}
-                                                    />,
-                                                ]}
-                                            />
-                                        </Text>
-                                    </Show.Eu>
-                                    <Show.NonEU>
-                                        <Text mt="2rem" max_width="58.8rem">
-                                            <Localize
-                                                translate_text="For fair resolution of any complaints, please speak to us using <0>chat</0>."
-                                                components={[
-                                                    <TextLink
-                                                        key={0}
-                                                        color="red"
-                                                        onClick={() => {
-                                                            LC_API.open_chat_window()
-                                                        }}
-                                                    />,
-                                                ]}
-                                            />
-                                        </Text>
-                                    </Show.NonEU>
-                                </div>
-                            )}
+                            <FinancialCommission />
                         </CssGridColumn>
                     </CssGrid>
                 </SmallContainer>
